+++ examples/index.js
@@ -1,5 +1,4 @@
 import ansis from '../src/index.js';
-import chalk from 'chalk';
 
 const log = console.log;
 const style = 'green';
@@ -176,7 +175,9 @@ log('BG HEX 6 color: ', ansis.bgHex('#FFBBAA').black('foo'));
 log('FG RGB color: ', ansis.rgb(100, 150, 200)('foo'));
 log('BG RGB color: ', ansis.bgRgb(100, 150, 200).black('foo'));
 log('FG ANSI 256 color: ', ansis.ansi256(96)('foo'));
+log('FG ANSI 256 color: ', ansis.fg(96)('foo alias'));
 log('BG ANSI 256 color: ', ansis.black.bgAnsi256(106)('foo'));
+log('BG ANSI 256 color: ', ansis.black.bg(106)('foo alias'));
 log();
 
 // background color